move `UserDataAutoSyncService`
