+++ src/vs/code/node/sharedProcess/sharedProcessMain.ts
@@ -109,6 +109,7 @@ import { ExtensionsProfileScannerService } from 'vs/platform/extensionManagement
 import { RequestChannelClient } from 'vs/platform/request/common/requestIpc';
 import { ExtensionRecommendationNotificationServiceChannelClient } from 'vs/platform/extensionRecommendations/common/extensionRecommendationsIpc';
 import { INativeHostService } from 'vs/platform/native/common/native';
+import { UserDataAutoSyncService } from 'vs/platform/userDataSync/node/userDataAutoSyncService';
 
 /* eslint-disable local/code-layering, local/code-import-patterns */
 // TODO@bpasero layer is not allowed in utility process
@@ -116,7 +117,6 @@ import { ExtensionTipsService } from 'vs/platform/extensionManagement/electron-s
 import { MessagePortMainProcessService } from 'vs/platform/ipc/electron-browser/mainProcessService';
 import { IMainProcessService } from 'vs/platform/ipc/electron-sandbox/services';
 import { NativeStorageService } from 'vs/platform/storage/electron-sandbox/storageService';
-import { UserDataAutoSyncService } from 'vs/platform/userDataSync/electron-sandbox/userDataAutoSyncService';
 import { UserDataProfileStorageService } from 'vs/platform/userDataProfile/electron-sandbox/userDataProfileStorageService';
 
 // TODO@bpasero remove these once utility process is the only way
