+++ src/vs/platform/userDataSync/node/userDataAutoSyncService.ts
